split and simplify further
